+++ frontend/src/app/components/products/products-all/products-all.component.html
@@ -1,3 +1,26 @@
-<p>
-  products-all works!
-</p>
+<ng-container *ngIf="products">
+      <a *ngFor="let product of products; index as i" [routerLink]="['/new/single-product/',product.idProduct]">
+        <div class="col-xl-4 col-lg-4 col-md-6">
+          <div class="single-product mb-60">
+            <div class="product-img">
+              <img [src]="imagesProducts[i]" alt={{product.productName}}>
+              <div class="new-product">
+                <span>{{product.category}}</span>
+              </div>
+            </div>
+            <div class="product-caption">
+              <div class="product-ratting">
+                <ng-template ngFor let-number [ngForOf]="[1,2,3,4,5]">
+                  <i *ngIf="ratingProducts[i]>=number" class="far fa-star"></i>
+                  <i *ngIf="ratingProducts[i]<number" class="far fa-star low-star"></i>
+                </ng-template>
+              </div>
+              <h4><a href="#">{{product.productName}}</a></h4>
+              <p>${{product.price}}</p>
+            </div>
+          </div>
+        </div>
+      </a>
+    </ng-container>
+    <h4 *ngIf="!products">No products yet.</h4>
+  <button type="button" id="bLoadMoreCategory" class="btn header-btn" *ngIf="loadMoreProductsVisible" (click)="getProductsPage()">Load More</button>
