+++ frontend/src/app/components/products/products-rating/products-rating.component.html
@@ -1,3 +1,26 @@
-<p>
-  products-rating works!
-</p>
+<ng-container *ngIf="productsRating">
+  <a *ngFor="let product of productsRating; index as i" [routerLink]="['/new/single-product/',product.idProduct]">
+    <div class="col-xl-4 col-lg-4 col-md-6">
+      <div class="single-product mb-60">
+        <div class="product-img">
+          <img [src]="imagesProductsRating[i]" alt={{product.productName}}>
+          <div class="new-product">
+            <span>{{product.category}}</span>
+          </div>
+        </div>
+        <div class="product-caption">
+          <div class="product-ratting">
+            <ng-template ngFor let-number [ngForOf]="[1,2,3,4,5]">
+              <i *ngIf="ratingProductsRating[i]>=number" class="far fa-star"></i>
+              <i *ngIf="ratingProductsRating[i]<number" class="far fa-star low-star"></i>
+            </ng-template>
+          </div>
+          <h4><a href="#">{{product.productName}}</a></h4>
+          <p>${{product.price}}</p>
+        </div>
+      </div>
+    </div>
+  </a>
+</ng-container>
+<h4 *ngIf="!productsRating">No products yet.</h4>
+<button type="button" id="bLoadMoreCategory" class="btn header-btn" *ngIf="loadMoreRatingVisible" (click)="getProductsRatingPage()">Load More</button>
