+++ frontend/src/app/components/profile/profile.component.html
@@ -36,10 +36,10 @@
                 [routerLink]="['/new/profile/profile-newproduct']" role="tab" aria-controls="nav-profile"
                 aria-selected="false" style="color:grey">Add Product</a>
               <a class="nav-item nav-link" id="nav-profile-tab" data-toggle="tab"
-                [routerLink]="['/new/profile/profile-sales']" role="tab" aria-controls="nav-profile"
+                [routerLink]="['/new/profile/profile-shopping']" role="tab" aria-controls="nav-profile"
                 aria-selected="false" style="color:grey">Your Shopping</a>
               <a class="nav-item nav-link" id="nav-profile-tab" data-toggle="tab"
-                [routerLink]="['/new/profile/profile-shopping']" role="tab" aria-controls="nav-profile"
+                [routerLink]="['/new/profile/profile-sales']" role="tab" aria-controls="nav-profile"
                 aria-selected="false" style="color:grey">Your Sales</a>
             </div>
           </nav>
